Реализована поддержка 48-bit LBA в ATA PIO/DMA (диски >128GB), 64-bit BAR и рекурсивный bridge enumeration в PCI (устройства за мостами и MMIO >4GB), и исправлены xHCI/AHCI для использования полных 64-bit адресов.
Что исправлено
1. ATA PIO: 48-bit LBA (ata.h, ata.c)
ATADevice.total_sectors — uint32_t → uint64_t (поддержка дисков до 128 PB)
ATADevice.lba48_supported — новое поле, детектится из IDENTIFY word 83 bit 10
ata_read_sectors / ata_write_sectors — параметр lba теперь uint64_t, count — uint16_t
При lba > 0x0FFFFFFF или count > 256 автоматически используются команды READ/WRITE SECTORS EXT (0x24/0x34) с полным HOB-протоколом
ata_flush_cache — использует FLUSH CACHE EXT (0xEA) для LBA48 устройств
2. ATA DMA: 48-bit LBA (ata_dma.h, ata_dma.c)
ata_dma_request_t.lba — uint32_t → uint64_t
Добавлены ATA_CMD_READ_DMA_EXT (0x25) / ATA_CMD_WRITE_DMA_EXT (0x35)
ata_dma_send_command автоматически выбирает 28-bit или 48-bit LBA в зависимости от устройства и адреса
3. AHCI Sync: 48-bit LBA passthrough (ahci_sync.h, ahci_sync.c)
Параметр lba изменён на uint64_t для сквозной передачи в AHCI NCQ (который уже работал с 48-bit)
4. PCI: 64-bit BAR + Bridge enumeration (pci.h, pci.c)
pci_read_bar64() — новая функция: читает BAR[n] и BAR[n+1] для 64-bit MMIO, проверяет type bits [2:1]
Рекурсивный PCI bridge scan — pci_scan_bus() рекурсивно следует за PCI-to-PCI bridges (header type 1) через PCI_BRIDGE_SECONDARY_BUS
Multi-domain host bridge — проверяет MF-бит host bridge для поддержки нескольких PCI доменов
Добавлены defines: PCI_HEADER_TYPE_BRIDGE, PCI_BRIDGE_*, PCI_BAR_TYPE_64BIT и др.
5. xHCI: 64-bit BAR (xhci.c)
Заменён pci_read_bar() на pci_read_bar64() — xHCI контроллеры с MMIO выше 4GB теперь корректно маппятся
6. AHCI: 64-bit BAR (ahci.c)
Заменён ручной 64-bit BAR парсинг (который имел баг с BAR6 > 5) на pci_read_bar64()
7. ACPI Shutdown (acpi_shutdown.c)
Убрана зависимость от #ifdef CONFIG_ACPI_FALLBACK_QEMU — QEMU fallback всегда доступен как last resort после неудачного PM1
Что осталось на будущее (отдельные этапы)
Задача	Сложность
SMP/AP startup (multi-core)	Большая — trampoline, per-CPU, IPI
PCIe ECAM (MCFG)	Средняя — нужен MCFG parsing + MMIO config space
Полный AML interpreter	Очень большая — целый проект
NCQ для legacy IDE	Невозможно — PIIX4 не поддерживает (только AHCI)

+++ src/kernel/drivers/disk/ahci_sync.c
@@ -6,7 +6,7 @@
 #include "klib.h"
 #include "atomics.h"
 
-int ahci_read_sectors_sync(uint8_t port, uint32_t lba,
+int ahci_read_sectors_sync(uint8_t port, uint64_t lba,
                            uint16_t sector_count, void* buffer) {
     if (!buffer || sector_count == 0) return -1;
 
@@ -92,7 +92,7 @@ int ahci_read_sectors_sync(uint8_t port, uint32_t lba,
     return -1;
 }
 
-int ahci_write_sectors_sync(uint8_t port, uint32_t lba,
+int ahci_write_sectors_sync(uint8_t port, uint64_t lba,
                             uint16_t sector_count, const void* buffer) {
     if (!buffer || sector_count == 0) return -1;
